Removing jQuery Dependency
The only pieces that required jQuery seemed to be the POST and a simple call for parameter encoding in a URI scheme. I rewrote those in native JS so that it is more portable of an example and something that people can adapt if they need whether or not they use jQuery.